zbus: use http between servers and not plain redis
Original implementation was using redis everywhere to send and receive
message, local and remote side. This way makes communication easy and
the same everywhere but this require to have redis listening on public
interface, which is a huge security issue.

In order to avoid redis listening externally, servers won't use redis
directly to talk to each other but will use a simple HTTP server and
talk directly from one to the other like this.

The HTTP server only listen on POST '/zbus-remote' and '/zbus-reply'.

A simple json verification is done, then json is forwarded to local
redis (queue 'msgbus.system.remote' or 'msgbus.system.reply').

Workflow is the same as before after that. Now redis can only listen on
local interfaces and process local requests.